Moved hosts.py under protocol submodule and some config organized
