+++ pyworkflow/utils/process.py
@@ -96,17 +96,6 @@ def buildRunCommand(programname, params, numberOfMpi, hostConfig=None,
         }
 
 
-def loadHostConfig(host='localhost'):
-    """ This function will load the execution host configuration.
-    In there will be information to know how to launch MPI processes
-    and how to submit jobs to the queue system if exists.
-    """
-    from pyworkflow.hosts import HostMapper
-    from pyworkflow import HOME
-    mapper = HostMapper(os.path.join(HOME, '..', 'settings'))
-    return mapper.selectByLabel(host)
-
-
 def killWithChilds(pid):
     """ Kill the process with given pid and all children processes.
     Params:
@@ -124,11 +113,12 @@ def killWithChilds(pid):
     else:
         proc.kill()
 
+
 def isProcessAlive(pid):
     import psutil
     try:
         proc = psutil.Process(pid)
         return proc.is_running()
-    except psutil.NoSuchProcess, e:
+    except psutil.NoSuchProcess:
         return False
